game many changes in code: removed one statistic class & added changes in level statistics class
