+++ src/main/java/com/waterworld/level/Point.java
@@ -23,15 +23,15 @@ public class Point {
         if (level.equals("level one")) {
             g.setColor(new Color(225, 74, 83));
             g.setFont(new Font("Showcard Gothic", Font.PLAIN, 28));
-            g.drawString(score / 10 + String.valueOf(score % 10) + "/50", (GameEngine.WIDTH / 16), 36);
+            g.drawString(score / 10 + String.valueOf(score % 10) + "/25", (GameEngine.WIDTH / 16), 36);
         } else if (level.equals("level two")) {
             g.setColor(new Color(225, 74, 83));
             g.setFont(new Font("Showcard Gothic", Font.PLAIN, 28));
-            g.drawString(score / 10 + String.valueOf(score % 10) + "/100", (GameEngine.WIDTH / 16), 36);
+            g.drawString(score / 10 + String.valueOf(score % 10) + "/50", (GameEngine.WIDTH / 16), 36);
         } else if (level.equals("level three")) {
             g.setColor(new Color(225, 74, 83));
             g.setFont(new Font("Showcard Gothic", Font.PLAIN, 28));
-            g.drawString(score / 10 + String.valueOf(score % 10) + "/200", (GameEngine.WIDTH / 16), 36);
+            g.drawString(score / 10 + String.valueOf(score % 10) + "/100", (GameEngine.WIDTH / 16), 36);
         }
         g.setColor(new Color(225, 74, 83));
         g.setFont(new Font("Showcard Gothic", Font.PLAIN, 28));
